+++ Translate.h
@@ -7,13 +7,13 @@ public:
     translate(Hittable *p, const Vec3& displacement)
             : ptr(p), offset(displacement) {}
     virtual bool hit(
-            const Ray& r, float t_min, float t_max, hit_record& rec) const;
+            const Ray& r, double t_min, double t_max, hit_record& rec) const;
     virtual bool bounding_box(float t0, float t1, AABB& box) const;
     Hittable *ptr;
     Vec3 offset;
 };
 
-bool translate::hit(const Ray& r, float t_min, float t_max, hit_record& rec) const {
+bool translate::hit(const Ray& r, double t_min, double t_max, hit_record& rec) const {
     Ray moved_r(r.origin() - offset, r.direction());
     if (ptr->hit(moved_r, t_min, t_max, rec)) {
         rec.p += offset;
